+++ src/privateer/backup.py
@@ -1,6 +1,6 @@
-import glob
 import os
-import tempfile
+import random
+import string
 from typing import List
 
 import docker
@@ -8,21 +8,33 @@ from fabric import Connection
 from invoke import UnexpectedExit
 
 from privateer.config import PrivateerHost, PrivateerTarget
-from privateer.docker_helpers import DockerClient
+from privateer.docker_helpers import DockerClient, containers_matching
 
 
-def backup(host: PrivateerHost, targets: List[PrivateerTarget]):
-    check_host_path(host)
+def get_mounts(host):
     mounts = [
         docker.types.Mount("/etc/timezone", "/etc/timezone", type="bind"),
         docker.types.Mount("/etc/localtime", "/etc/localtime", type="bind"),
     ]
     if host.host_type == "remote":
-        env = {"SSH_HOST_NAME": host.hostname, "SSH_REMOTE_PATH": host.path, "SSH_USER": host.user}
         mounts.append(docker.types.Mount("/root/.ssh", os.path.expanduser("~/.ssh"), type="bind"))
     else:
         mounts.append(docker.types.Mount("/archive", host.path, type="bind"))
+    return mounts
+
+
+def get_env(host):
+    if host.host_type == "remote":
+        env = {"SSH_HOST_NAME": host.hostname, "SSH_REMOTE_PATH": host.path, "SSH_USER": host.user}
+    else:
         env = {}
+    return env
+
+
+def backup(host: PrivateerHost, targets: List[PrivateerTarget]):
+    check_host_path(host)
+    mounts = get_mounts(host)
+    env = get_env(host)
 
     for t in targets:
         mounts.append(docker.types.Mount(f"/{t.name}", t.name))
@@ -38,40 +50,46 @@ def backup(host: PrivateerHost, targets: List[PrivateerTarget]):
     return True
 
 
-def restore(host: PrivateerHost, targets: List[PrivateerTarget]):
-    success = []
-    if host.host_type == "local":
-        for t in targets:
-            path = get_most_recent_backup(host.path, t.name)
-            if path is None:
-                msg = f"No backups found. Not restoring {t.name}"
-                print(msg)
-            else:
-                untar_volume(t, path)
-                success.append(t.name)
-                print(f"Restored {path} to {t.name}")
-    else:
-        with Connection(host=host.hostname, user=host.user, port=host.port) as c:
-            with tempfile.TemporaryDirectory() as local_backup_path:
-                for t in targets:
-                    res = c.run(f"ls -t {host.path}/{t.name}-*.tar.gz | head -1", in_stream=False, pty=True)
-                    if res.ok:
-                        file = res.stdout.strip()
-                        c.get(file, f"{local_backup_path}/")
-                        untar_volume(t, f"{local_backup_path}/{os.path.basename(file)}")
-                        success.append(t.name)
-                        print(f"Restored {file} to {t.name}")
-                    else:
-                        msg = f"No backups found. Not restoring {t.name}"
-                        print(msg)
-    return success
+def generate_backup_config(target: PrivateerTarget):
+    if len(target.schedules) == 0:
+        msg = f"No backup schedules defined for target {target.name}"
+        raise Exception(msg)
+    if not os.path.exists("offen"):
+        os.mkdir("offen")
+    filenames = []
+    for s in target.schedules:
+        filename = f"offen/{target.name}-{s.name}.conf"
+        filenames.append(filename)
+        with open(filename, "w") as f:
+            f.write(f'BACKUP_SOURCES="/backup/{target.name}"\n')
+            f.write(f'BACKUP_FILENAME="{target.name}-{s.name}-%Y-%m-%dT%H-%M-%S.tar.gz"\n')
+            f.write(f'BACKUP_PRUNING_PREFIX="{target.name}-{s.name}-"\n')
+            f.write(f'BACKUP_CRON_EXPRESSION="{s.schedule}"\n')
+            if s.retention_days is not None:
+                f.write(f'BACKUP_RETENTION_DAYS="{s.retention_days}"\n')
+    return filenames
 
 
-def get_most_recent_backup(local_path, target_name):
-    backups = glob.glob(f"{local_path}/{target_name}*.tar.gz")
-    if len(backups) == 0:
-        return None
-    return max(backups, key=os.path.getctime)
+def schedule_backups(host: PrivateerHost, targets: List[PrivateerTarget]):
+    check_host_path(host)
+    mounts = get_mounts(host)
+    env = get_env(host)
+    for t in targets:
+        mounts.append(docker.types.Mount(f"/backup/{t.name}", t.name))
+        generate_backup_config(t)
+    path = os.path.abspath("offen")
+    mounts.append(docker.types.Mount("/etc/dockervolumebackup/conf.d", path, type="bind"))
+    name = f"privateer_{''.join(random.choices(string.ascii_letters, k=6))}"  # noqa: S311
+    with DockerClient() as cl:
+        cl.containers.run(
+            "offen/docker-volume-backup:v2", name=name, mounts=mounts, environment=env, detach=True, remove=True
+        )
+    return True
+
+
+def cancel_scheduled_backups():
+    running = containers_matching("privateer")
+    [r.stop() for r in running]
 
 
 def check_host_path(host: PrivateerHost):
@@ -86,16 +104,3 @@ def check_host_path(host: PrivateerHost):
             except UnexpectedExit as err:
                 msg = f"Host path '{host.path}' does not exist. Either make directory or fix config."
                 raise Exception(msg) from err
-
-
-def untar_volume(target: PrivateerTarget, backup_path):
-    volume_mount = docker.types.Mount("/data", target.name)
-    backup_mount = docker.types.Mount(f"/backup/{backup_path}", backup_path, type="bind")
-    with DockerClient() as cl:
-        cl.containers.run(
-            "ubuntu",
-            remove=True,
-            mounts=[volume_mount, backup_mount],
-            command=["tar", "xvf", f"/backup/{backup_path}", "-C", "/data", "--strip-components=1"],
-        )
-    return True
